Renamed meta optimizer to SGO optimizer.
