+++ Src/ILGPU.Algorithms/Optimization/CPU/SGOOptimizer.Scalar.cs
@@ -1,104 +1,87 @@
 // ---------------------------------------------------------------------------------------
 //                                   ILGPU Algorithms
-//                           Copyright (c) 2023 ILGPU Project
+//                        Copyright (c) 2023-2024 ILGPU Project
 //                                    www.ilgpu.net
 //
-// File: MetaOptimizer.Vectorized.cs
+// File: SGOOptimizer.Scalar.cs
 //
 // This file is part of ILGPU and is distributed under the University of Illinois Open
 // Source License. See LICENSE.txt for details.
 // ---------------------------------------------------------------------------------------
 
-using System.Numerics;
 using System.Runtime.CompilerServices;
 
 #if NET7_0_OR_GREATER
 
 namespace ILGPU.Algorithms.Optimization.CPU
 {
-    partial class MetaOptimizer<T, TEvalType>
+    partial class SGOOptimizer<T, TEvalType>
     {
         /// <summary>
-        /// A vectorized processor using SIMD operations.
+        /// A scalar processor using default ALUs.
         /// </summary>
-        private readonly struct VectorizedProcessor :
-            IProcessor<VectorizedProcessor, Vector<T>>
+        private readonly struct ScalarProcessor : IProcessor<ScalarProcessor, T>
         {
             /// <summary>
-            /// Creates a new vectorized processor.
+            /// Creates a new scalar processor.
             /// </summary>
-            public static VectorizedProcessor New() => default;
+            public static ScalarProcessor New() => default;
 
             /// <summary>
-            /// Returns the vector length.
+            /// Returns 1;
             /// </summary>
-            public static int Length => Vector<int>.Count;
+            public static int Length => 1;
 
             /// <summary>
-            /// Clamps the given vector.
+            /// Clamps the given value.
             /// </summary>
             /// <param name="lower">The lower bounds part.</param>
             /// <param name="upper">The upper bounds part.</param>
-            /// <param name="value">The vector to clamp.</param>
+            /// <param name="value">The value to clamp.</param>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public Vector<T> Clamp(
-                Vector<T> lower,
-                Vector<T> upper,
-                Vector<T> value) =>
-                Vector.Min(Vector.Max(value, lower), upper);
+            public T Clamp(T lower, T upper, T value) =>
+                T.Clamp(value, lower, upper);
 
             /// <summary>
             /// Resets the given data view.
             /// </summary>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public void Reset(out Vector<T> data) =>
-                data = new Vector<T>(T.Zero);
+            public void Reset(out T data) => data = T.Zero;
 
             /// <summary>
             /// Adds the given source to the target view.
             /// </summary>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public void Accumulate(ref Vector<T> target, Vector<T> source)
-            {
-                var accumulated = source + target;
-                target = accumulated;
-            }
+            public void Accumulate(ref T target, T source) =>
+                target += source;
 
             /// <summary>
             /// Computes the average by taking the given count into account.
             /// </summary>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public void ComputeAverage(ref Vector<T> target, T count)
-            {
-                var countValue = new Vector<T>(count);
-                var average = target / countValue;
-                target = average;
-            }
+            public void ComputeAverage(ref T target, T count) =>
+                target /= count;
 
             /// <summary>
-            /// Determines a newly sampled position using vectors.
+            /// Determines a newly sampled position using scalars.
             /// </summary>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public Vector<T> GetRandomPosition(
-                Vector<T> lower,
-                Vector<T> upper,
-                Vector<T> randomNumber)
+            public T GetRandomPosition(T lower, T upper, T randomNumber)
             {
                 // Interpolate between lower and upper bound
-                var lowerFactor = new Vector<T>(T.One) - randomNumber;
-                var lowerInfluence = lowerFactor * lower;
+                var lowerInfluence = (T.One - randomNumber) * lower;
                 var upperInfluence = randomNumber * upper;
                 return lowerInfluence + upperInfluence;
             }
 
             /// <summary>
-            /// Determines a newly sampled position using vectors.
+            /// Determines a newly sampled position using scalars.
             /// </summary>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public Vector<T> DetermineNewPosition(
-                Vector<T> position,
-                Vector<T> firstC,
-                Vector<T> secondC,
+            public T DetermineNewPosition(
+                T position,
+                T firstC,
+                T secondC,
                 T r1,
                 T r2,
                 T stepSize)
